Abstract out Hits
