+++ src/components/search/pages/SearchPage.js
@@ -1,29 +1,14 @@
 import React from 'react';
 import { InstantSearch, connectSearchBox, Index, Configure, connectHits } from 'react-instantsearch-dom';
-import { Text, ListChoice } from '@kiwicom/orbit-components/lib';
+import { Text } from '@kiwicom/orbit-components/lib';
 import SearchBar from '../SearchBar';
 import algoliasearch from 'algoliasearch/lite';
 import styled from 'styled-components';
-import { useRouter } from 'next/router';
+import Hits from '../modules/Hits';
 
 const searchClient = algoliasearch(process.env.ALGOLIA_APP_ID, process.env.ALGOLIA_ADMIN_KEY);
 
 const CustomSearchBox = connectSearchBox(SearchBar);
-
-const Hits = ({ hits, type }) => {
-  const router = useRouter();
-  const handleOnClick = (type, objectID) => {
-    router.push(`/${type}/${objectID}`);
-  };
-  return (
-    <>
-      {hits.map((hit) => (
-        <ListChoice description={hit.description} title={hit.title} onClick={() => handleOnClick(type, hit.objectID)} />
-      ))}
-    </>
-  );
-};
-
 const CustomHits = connectHits(Hits);
 
 const Container = styled.div`
